+++ test/unit/directives/mwlElementDimensions.spec.js
@@ -35,4 +35,11 @@ describe('mwlElementDimensions directive', function() {
     expect(scope.elementDimensions).to.eql({width: 100, height: 50});
   });
 
+  it('should update the element dimensions when the window is resized', function() {
+    element[0].style.width = '150px';
+    element[0].style.height = '20px';
+    $window.dispatchEvent(new $window.Event('resize'));
+    expect(scope.elementDimensions).to.eql({width: 150, height: 20});
+  });
+
 });